Fix examples webhook
