+++ packages/react-openapi/src/OpenAPIResponseExample.tsx
@@ -1,15 +1,11 @@
 import type { OpenAPIV3 } from '@gitbook/openapi-parser';
 import { Markdown } from './Markdown';
-import {
-    OpenAPIEmptyExample,
-    OpenAPIExample,
-    getExampleFromReference,
-    getExamplesFromMediaTypeObject,
-} from './OpenAPIExample';
+import { OpenAPIEmptyExample, OpenAPIExample } from './OpenAPIExample';
+import { OpenAPIMediaTypeContent } from './OpenAPIMediaType';
 import { OpenAPIResponseExampleContent } from './OpenAPIResponseExampleContent';
-import { OpenAPIResponseMediaTypeContent } from './OpenAPIResponseMediaType';
 import type { OpenAPIContext, OpenAPIOperationData, OpenAPIWebhookData } from './types';
-import { getStatusCodeDefaultLabel, getSyntaxFromMediaType } from './utils';
+import { getExampleFromReference, getExamples } from './util/example';
+import { createStateKey, getStatusCodeDefaultLabel } from './utils';
 import { checkIsReference, resolveDescription } from './utils';
 
 /**
@@ -127,30 +123,10 @@ function OpenAPIResponse(props: {
     });
 
     return (
-        <OpenAPIResponseMediaTypeContent
+        <OpenAPIMediaTypeContent
             selectIcon={context.icons.chevronDown}
-            blockKey={context.blockKey}
+            stateKey={createStateKey('response-media-types', context.blockKey)}
             items={tabs}
         />
     );
-}
-
-function getExamples(props: {
-    mediaTypeObject: OpenAPIV3.MediaTypeObject;
-    mediaType: string;
-    context: OpenAPIContext;
-}) {
-    const { mediaTypeObject, mediaType } = props;
-    const examples = getExamplesFromMediaTypeObject({ mediaTypeObject, mediaType });
-    const syntax = getSyntaxFromMediaType(mediaType);
-
-    return examples.map((example) => {
-        return {
-            key: example.key,
-            label: example.example.summary || example.key,
-            body: (
-                <OpenAPIExample example={example.example} context={props.context} syntax={syntax} />
-            ),
-        };
-    });
 }
